multiple objects rendering

+++ src/ray.h
@@ -14,7 +14,7 @@ public:
 
   vec3 origin() const    { return A; }
   vec3 direction() const { return B; }
-  vec3 point_at_parameter(float t) const { return A + t * B; }
+  vec3 point_at_parameter(double t) const { return A + t * B; }
 
   vec3 A;
   vec3 B;
